Implement view patterns
