+++ data/examples/declaration/value/function/view-pattern-out.hs
@@ -1,0 +1,14 @@
+{-# LANGUAGE ViewPatterns #-}
+example f (f -> 4) = True
+
+f (t -> Nothing) = "Nothing"
+f (t -> Just _) = "Just"
+
+g ((f, _), f -> 4) = True
+
+multiline
+  ( t ->
+      Foo
+        bar
+        baz
+  ) = True
