+++ ormolu.cabal
@@ -61,7 +61,6 @@ library
                     , Ormolu.Printer.Meat.Common
                     , Ormolu.Printer.Meat.Declaration
                     , Ormolu.Printer.Meat.Declaration.Data
-                    , Ormolu.Printer.Meat.Declaration.Pat
                     , Ormolu.Printer.Meat.Declaration.Signature
                     , Ormolu.Printer.Meat.Declaration.Type
                     , Ormolu.Printer.Meat.Declaration.TypeFamily
